pyproject.toml
[project]
name = "qadapt"
version = "0.1.0"
description = "Reinforcement learning agents for quantum device tuning"
requires-python = ">=3.11,<3.12"
dependencies = [
    # QArray
    "qarray==1.6.0",
    "qarray-latched @ git+ssh://git@github.com/b-vanstraaten/qarray-latched.git@fcc472276f27e7633bb3aafc6f0d6c92966875d7",
    # Core ML/RL - pinned to match working conda env
    "torch==2.8.0",
    "torchvision==0.23.0",
    "ray[rllib]==2.51",
    "wandb[media]",
    "gymnasium==1.1.1",
    "jax[cuda12]==0.7.2",
    "jaxlib==0.7.2",
    "dynamiqs==0.3.4",  # SuperSims All-XY transmon simulator
    # Data and visualization - pinned versions
    "numpy==2.2.4",
    "matplotlib==3.10.3",
    "pyyaml==6.0.2",
    "scipy==1.16.2",
    "scikit-learn==1.7.1",
    "scienceplots",
    "gputil>=1.4.0",
    "modal>=1.3.0.post1",
    "stable-baselines3>=2.7.1",
    "cleanrl>=0.4.8",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]
facmac = [
    "sacred",
    "jsonpickle",
    "tensorboard-logger",
    "gym==0.10.8",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src", "benchmarks/MARL"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs | \.git | \.mypy_cache | \.venv | build | dist | wandb
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten later
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ray.*",
    "wandb.*",
    "matplotlib.*",
    "scipy.*",
    "qarray.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.uv]
extra-index-url = [
    "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html",
    "https://download.pytorch.org/whl/cu128",
]
# dynamiqs 0.3.4's metadata caps jax<0.7, but it actually works with jax 0.7.2
# at runtime (verified on supersims-qadapt). Override the upper bound so uv lets
# both coexist. Drop this once a dynamiqs release relaxes the pin.
override-dependencies = [
    "jax>=0.4.36",
]
